+++ packages/jsx/src/index.test.tsx
@@ -16,7 +16,7 @@ type SetupFn = (
 ) => void
 
 const setup = (fn: SetupFn) => async () => {
-  const ctx = createTestCtx()
+  const ctx = createTestCtx({ restrictMultipleContexts: false })
   const { h, hf, mount } = reatomJsx(ctx, window)
 
   const parent = window.document.createElement('div')
@@ -42,10 +42,10 @@ it(
   setup((ctx, h, hf, mount, parent) => {
     const element = <div id="some-id">Hello, world!</div>
 
-    expect(element.tagName).toEqual('DIV')
-    expect(element.id).toEqual('some-id')
-    expect(element.childNodes.length).toEqual(1)
-    expect(element.textContent).toEqual('Hello, world!')
+    expect(element.tagName).toBe('DIV')
+    expect(element.id).toBe('some-id')
+    expect(element.childNodes.length).toBe(1)
+    expect(element.textContent).toBe('Hello, world!')
   }),
 )
 
@@ -60,19 +60,19 @@ it(
 
     mount(parent, element)
 
-    expect(element.id).toEqual('val')
+    expect(element.id).toBe('val')
     // @ts-expect-error `dunno` can't be inferred
-    expect(element.prp).toEqual('prp')
-    expect(element.getAttribute('atr')).toEqual('atr')
+    expect(element.prp).toBe('prp')
+    expect(element.getAttribute('atr')).toBe('atr')
 
     val(ctx, 'val1')
     prp(ctx, 'prp1')
     atr(ctx, 'atr1')
 
-    expect(element.id).toEqual('val1')
+    expect(element.id).toBe('val1')
     // @ts-expect-error `dunno` can't be inferred
-    expect(element.prp, 'prp1')
-    expect(element.getAttribute('atr')).toEqual('atr1')
+    expect(element.prp).toBe('prp1')
+    expect(element.getAttribute('atr')).toBe('atr1')
   }),
 )
 
@@ -94,16 +94,16 @@ it(
 
     mount(parent, element)
 
-    expect(element.childNodes.length).toEqual(5)
-    expect(element.childNodes[2]?.textContent).toEqual('foo')
-    expect(element.childNodes[4]).toEqual(a)
+    expect(element.childNodes.length).toBe(7)
+    expect(element.childNodes[2]?.textContent).toBe('foo')
+    expect(element.childNodes[5]).toBe(a)
 
     val(ctx, 'bar')
-    expect(element.childNodes[2]?.textContent).toEqual('bar')
+    expect(element.childNodes[2]?.textContent).toBe('bar')
 
-    expect(element.childNodes[4]).toEqual(a)
+    expect(element.childNodes[5]).toBe(a)
     route(ctx, 'b')
-    expect(element.childNodes[4]).toEqual(b)
+    expect(element.childNodes[5]).toBe(b)
   }),
 )
 
@@ -116,14 +116,14 @@ it(
 
     mount(parent, element)
 
-    expect(element.childNodes.length).toEqual(2)
+    expect(element.childNodes.length).toBe(3)
 
     children(ctx, <div>Hello, world!</div>)
-    expect(element.childNodes[1]?.textContent).toEqual('Hello, world!')
+    expect(element.childNodes[1]?.textContent).toBe('Hello, world!')
 
     const inner = <span>inner</span>
     children(ctx, <div>{inner}</div>)
-    expect(element.childNodes[1]?.childNodes[0]).toEqual(inner)
+    expect(element.childNodes[1]?.childNodes[0]).toBe(inner)
 
     const before = atom('before', 'before')
     const after = atom('after', 'after')
@@ -135,10 +135,10 @@ it(
         {after}
       </div>,
     )
-    expect((element as HTMLDivElement).textContent).toEqual('beforeinnerafter')
+    expect((element as HTMLDivElement).innerText).toBe('beforeinnerafter')
 
     before(ctx, 'before...')
-    expect((element as HTMLDivElement).textContent).toEqual('before...innerafter')
+    expect((element as HTMLDivElement).innerText).toBe('before...innerafter')
   }),
 )
 
@@ -156,12 +156,53 @@ it(
 
     mount(parent, element)
 
-    expect(element.id).toEqual('1')
-    expect(element.getAttribute('b')).toEqual('2')
-    expect(clickTrack.calls.length).toEqual(0)
+    expect(element.id).toBe('1')
+    expect(element.getAttribute('b')).toBe('2')
+    expect(clickTrack.calls.length).toBe(0)
     // @ts-expect-error
     element.click()
-    expect(clickTrack.calls.length).toEqual(1)
+    expect(clickTrack.calls.length).toBe(1)
+  }),
+)
+
+it(
+  'multiple renden shared element',
+  setup(async (ctx, h, hf, mount, parent) => {
+    const valueAtom = atom('abc', 'value')
+    const element = <p>{valueAtom}</p>
+
+    const Component = () => (
+      <>
+        <div id="1">{element}</div>
+        <div id="2">{element}</div>
+      </>
+    )
+
+    const childAtom = atom<JSX.Element | undefined>(<Component></Component>, 'child')
+    const app = <div>{childAtom}</div>
+
+    mount(parent, app)
+    await sleep()
+    expect(app.innerHTML).toBe(
+      '<!--child--><!----><div id="1"></div><div id="2"><p><!--value-->abc<!--value--></p></div><!----><!--child-->',
+    )
+
+    valueAtom(ctx, 'def')
+    await sleep()
+    expect(app.innerHTML).toBe(
+      '<!--child--><!----><div id="1"></div><div id="2"><p><!--value-->def<!--value--></p></div><!----><!--child-->',
+    )
+
+    childAtom(ctx, undefined)
+    await sleep()
+    expect(app.innerHTML).toBe('<!--child--><!--child-->')
+
+    childAtom(ctx, <Component></Component>)
+    valueAtom(ctx, 'ghi')
+    await sleep()
+    expect(app.innerHTML).toBe(
+      '<!--child--><!----><div id="1"></div><div id="2"><p><!--value-->def<!--value--></p></div><!----><!--child-->',
+    )
   }),
 )
 
@@ -178,9 +219,8 @@ it(
     )
     mount(parent, child)
 
-    expect(parent.childNodes.length).toBe(2)
-    expect(parent.childNodes[0]?.textContent).toBe('foo')
-    expect(parent.childNodes[1]?.textContent).toBe('bar')
+    expect(parent.childNodes.length).toBe(6)
+    expect(parent.textContent).toBe('foobar')
   }),
 )
 
@@ -199,12 +239,12 @@ it(
 
     mount(parent, element)
 
-    expect(element.childNodes.length).toEqual(3)
-    expect(element.textContent).toEqual('1')
+    expect(element.childNodes.length).toBe(6)
+    expect(element.textContent).toBe('1')
 
     n(ctx, 2)
-    expect(element.childNodes.length).toEqual(4)
-    expect(element.textContent).toEqual('12')
+    expect(element.childNodes.length).toBe(7)
+    expect(element.textContent).toBe('12')
   }),
 )
 
@@ -218,15 +258,15 @@ it(
 
     mount(parent, <div>{jsxList}</div>)
 
-    expect(parent.textContent).toEqual('12')
-    expect(isConnected(ctx, one))
-    expect(isConnected(ctx, two))
+    expect(parent.innerText).toBe('12')
+    expect(isConnected(ctx, one)).toBe(true)
+    expect(isConnected(ctx, two)).toBe(true)
 
     list.swap(ctx, one, two)
-    expect(parent.textContent).toEqual('21')
+    expect(parent.innerText).toBe('21')
 
     list.remove(ctx, two)
-    expect(parent.textContent).toEqual('1')
+    expect(parent.innerText).toBe('1')
     await sleep()
     expect(isConnected(ctx, one)).toBe(true)
     expect(isConnected(ctx, two)).toBe(false)
@@ -250,8 +290,9 @@ it(
       </div>
     )
 
-    expect(element.childNodes.length).toEqual(2)
-    expect(element.textContent).toEqual('')
+    expect(element.childNodes.length).toBe(4)
+    expect(element.innerHTML).toBe('<!--true--><!--true--><!--false--><!--false-->')
+    expect(element.textContent).toBe('')
   }),
 )
 
@@ -268,8 +309,9 @@ it(
       </div>
     )
 
-    expect(element.childNodes.length).toEqual(1)
-    expect(element.textContent).toEqual('')
+    expect(element.childNodes.length).toBe(2)
+    expect(element.innerHTML).toBe('<!--null--><!--null-->')
+    expect(element.textContent).toBe('')
   }),
 )
 
@@ -286,8 +328,9 @@ it(
       </div>
     )
 
-    expect(element.childNodes.length).toEqual(1)
-    expect(element.textContent).toEqual('')
+    expect(element.childNodes.length).toBe(2)
+    expect(element.innerHTML).toBe('<!--undefined--><!--undefined-->')
+    expect(element.textContent).toBe('')
   }),
 )
 
@@ -304,8 +347,9 @@ it(
       </div>
     )
 
-    expect(element.childNodes.length).toEqual(1)
-    expect(element.textContent).toEqual('')
+    expect(element.childNodes.length).toBe(2)
+    expect(element.innerHTML).toBe('<!--emptyString--><!--emptyString-->')
+    expect(element.textContent).toBe('')
   }),
 )
 
@@ -318,13 +362,13 @@ it(
 
     mount(parent, element)
 
-    expect(parent.childNodes.length).toEqual(1)
-    expect(parent.textContent).toEqual('')
+    expect(parent.childNodes.length).toBe(1)
+    expect(parent.textContent).toBe('')
 
     valueAtom(ctx, 123)
 
-    expect(parent.childNodes.length).toEqual(1)
-    expect(parent.textContent).toEqual('123')
+    expect(parent.childNodes.length).toBe(1)
+    expect(parent.textContent).toBe('123')
   }),
 )
 
@@ -335,7 +379,7 @@ it(
 
     const element = <div>{htmlAtom}</div>
 
-    expect(element.innerHTML).toEqual('<!--html--><div>div</div>')
+    expect(element.innerHTML).toBe('<!--html--><div>div</div><!--html-->')
   }),
 )
 
@@ -346,7 +390,7 @@ it(
 
     const element = <div>{svgAtom}</div>
 
-    expect(element.innerHTML).toEqual('<!--svg--><svg>svg</svg>')
+    expect(element.innerHTML).toBe('<!--svg--><svg>svg</svg><!--svg-->')
   }),
 )
 
@@ -355,9 +399,9 @@ it(
   setup((ctx, h, hf, mount, parent) => {
     const Component = (props: JSX.HTMLAttributes) => <div {...props} />
 
-    expect(<Component />).instanceOf(window.HTMLElement)
-    expect(((<Component draggable="true" />) as HTMLElement).draggable).toEqual(true)
-    expect(((<Component>123</Component>) as HTMLElement).textContent).toEqual('123')
+    expect(<Component />).toBeInstanceOf(window.HTMLElement)
+    expect(((<Component draggable="true" />) as HTMLElement).draggable).toBe(true)
+    expect(((<Component>123</Component>) as HTMLElement).innerText).toBe('123')
   }),
 )
 
@@ -380,11 +424,11 @@ it(
     )
 
     mount(parent, component)
-    expect(ref).instanceOf(window.HTMLElement)
+    expect(ref).toBeInstanceOf(window.HTMLElement)
 
     parent.remove()
     await sleep()
-    expect(ref).toEqual(null)
+    expect(ref).toBe(null)
   }),
 )
 
@@ -407,12 +451,12 @@ it(
     )
 
     mount(parent, component)
-    expect(ref).instanceOf(window.HTMLElement)
+    expect(ref).toBeInstanceOf(window.HTMLElement)
     await sleep()
 
     ref!.remove()
     await sleep()
-    expect(ref).toEqual(null)
+    expect(ref).toBe(null)
   }),
 )
 
@@ -438,18 +482,18 @@ it(
     )
 
     mount(parent, component)
-    expect(ref).instanceOf(window.HTMLElement)
+    expect(ref).toBeInstanceOf(window.HTMLElement)
     await sleep()
 
     ref!.remove()
     await sleep()
-    expect(ref).toEqual(null)
+    expect(ref).toBe(null)
 
-    expect(mountArgs[0]).toEqual(ctx)
-    expect(mountArgs[1]).toEqual(component)
+    expect(mountArgs[0]).toBe(ctx)
+    expect(mountArgs[1]).toBe(component)
 
-    expect(unmountArgs[0]).toEqual(ctx)
-    expect(unmountArgs[1]).toEqual(component)
+    expect(unmountArgs[0]).toBe(ctx)
+    expect(unmountArgs[1]).toBe(component)
   }),
 )
 
@@ -470,17 +514,43 @@ it(
     )
 
     mount(parent, component)
-    expect(ref1).instanceOf(window.HTMLElement)
-    expect(ref2).instanceOf(window.HTMLElement)
+    expect(ref1).toBeInstanceOf(window.HTMLElement)
+    expect(ref2).toBeInstanceOf(window.HTMLElement)
     await sleep()
 
-    expect(ref1.className).toEqual(cls)
+    expect(ref1.className).toBe(cls)
     expect(ref1.dataset.reatom).toBeTruthy()
 
-    expect(ref2.className).toEqual(cls)
+    expect(ref2.className).toBe(cls)
     expect(ref2.dataset.reatom).toBeTruthy()
 
-    expect(ref1.dataset.reatom, ref2.dataset.reatom)
+    expect(ref1.dataset.reatom).toBe(ref2.dataset.reatom)
+  }),
+)
+
+it(
+  'css property generate class name',
+  setup(async (ctx, h, hf, mount, parent) => {
+    const css = 'color: red;'
+
+    const First = () => <div css={css}></div>
+    const Second = () => <div css={css}></div>
+
+    const first = <First></First>
+    const second = <Second></Second>
+
+    const component = (
+      <div>
+        {first}
+        {second}
+      </div>
+    )
+
+    mount(parent, component)
+    await sleep()
+
+    expect(first.dataset.reatom!.startsWith(First.name)).toBeTruthy()
+    expect(second.dataset.reatom!.startsWith(Second.name)).toBeTruthy()
   }),
 )
 
@@ -494,18 +564,18 @@ it(
     mount(parent, component)
     await sleep()
 
-    expect(component.style.getPropertyValue('--first-property'), 'red')
-    expect(component.style.getPropertyValue('--secondProperty'), 'red')
+    expect(component.style.getPropertyValue('--first-property')).toBe('red')
+    expect(component.style.getPropertyValue('--secondProperty')).toBe('red')
 
     colorAtom(ctx, 'green')
 
-    expect(component.style.getPropertyValue('--first-property'), 'green')
-    expect(component.style.getPropertyValue('--secondProperty'), 'green')
+    expect(component.style.getPropertyValue('--first-property')).toBe('green')
+    expect(component.style.getPropertyValue('--secondProperty')).toBe('green')
 
     colorAtom(ctx, undefined)
 
-    expect(component.style.getPropertyValue('--first-property'), '')
-    expect(component.style.getPropertyValue('--secondProperty'), '')
+    expect(component.style.getPropertyValue('--first-property')).toBe('')
+    expect(component.style.getPropertyValue('--secondProperty')).toBe('')
   }),
 )
 
@@ -527,20 +597,20 @@ it(
     mount(parent, component)
     await sleep()
 
-    expect(ref1.hasAttribute('class'))
-    expect(ref2.hasAttribute('class'))
+    expect(ref1.hasAttribute('class')).toBe(true)
+    expect(ref2.hasAttribute('class')).toBe(true)
 
     classAtom(ctx, 'cls')
-    expect(ref1.className).toEqual('cls')
-    expect(ref2.className).toEqual('cls')
-    expect(ref1.hasAttribute('class')).toBeTruthy()
-    expect(ref2.hasAttribute('class')).toBeTruthy()
+    expect(ref1.className).toBe('cls')
+    expect(ref2.className).toBe('cls')
+    expect(ref1.hasAttribute('class')).toBe(true)
+    expect(ref2.hasAttribute('class')).toBe(true)
 
     classAtom(ctx, undefined)
-    expect(ref1.className).toEqual('')
-    expect(ref2.className).toEqual('')
-    expect(!ref1.hasAttribute('class')).toBeTruthy()
-    expect(!ref2.hasAttribute('class')).toBeTruthy()
+    expect(ref1.className).toBe('')
+    expect(ref2.className).toBe('')
+    expect(ref1.hasAttribute('class')).toBe(false)
+    expect(ref2.hasAttribute('class')).toBe(false)
   }),
 )
 
@@ -608,71 +678,20 @@ it(
 
     mount(parent, component)
 
-    expect(firstEl.getAttribute('style')).toEqual('top: 0px; left: 0px;')
-    expect(secondEl.getAttribute('style')).toEqual('top: 0px; left: 0px;')
+    expect(firstEl.getAttribute('style')).toBe('top: 0px; left: 0px;')
+    expect(secondEl.getAttribute('style')).toBe('top: 0px; left: 0px;')
 
     styleTopAtom(ctx, undefined)
     styleBottomAtom(ctx, 0)
 
-    expect(firstEl.getAttribute('style')).toEqual('left: 0px; bottom: 0px;')
-    expect(secondEl.getAttribute('style')).toEqual('left: 0px; bottom: 0px;')
-  }),
-)
-
-it(
-  'render different atom children',
-  setup((ctx, h, hf, mount, parent) => {
-    const aChild = atom(<span>a</span>, 'aChild')
-    const a = atom(
-      <>
-        {aChild}
-        <span />
-      </>,
-      'a',
-    )
-
-    const bChild = atom(<div>b</div>, 'bChild')
-    const b = atom(
-      <>
-        <div />
-        {bChild}
-      </>,
-      'b',
-    )
-
-    const container = (
-      <main>
-        <>
-          {a}
-          {b}
-        </>
-      </main>
-    )
-
-    mount(parent, container)
-    expect(container.innerHTML).toEqual(
-      `<!--a--><!--aChild--><span>a</span><span></span><!--b--><div></div><!--bChild--><div>b</div>`,
-    )
-
-    aChild(ctx, <h1>A</h1>)
-    expect(container.innerHTML).toEqual(
-      `<!--a--><!--aChild--><h1>A</h1><span></span><!--b--><div></div><!--bChild--><div>b</div>`,
-    )
-
-    // @ts-expect-error
-    bChild(ctx, null)
-    expect(container.innerHTML).toEqual(`<!--a--><!--aChild--><h1>A</h1><span></span><!--b--><div></div><!--bChild-->`)
-
-    bChild(ctx, <div>B</div>)
-    expect(container.innerHTML).toEqual(
-      `<!--a--><!--aChild--><h1>A</h1><span></span><!--b--><div></div><!--bChild--><div>B</div>`,
-    )
+    expect(firstEl.getAttribute('style')).toBe('left: 0px; bottom: 0px;')
+    expect(secondEl.getAttribute('style')).toBe('left: 0px; bottom: 0px;')
   }),
 )
 
 it(
   'render atom fragments',
-  setup((ctx, h, hf, mount, parent) => {
+  setup(async (ctx, h, hf, mount, parent) => {
     const bool1Atom = atom(false)
     const bool2Atom = atom(false)
 
@@ -703,37 +722,46 @@ it(
       </div>
     )
 
-    const expect1 = '<p>0</p><!--1--><p>5</p>'
-    const expect2 = '<p>0</p><!--1--><p>1</p><!--2--><p>4</p><p>5</p>'
-    const expect3 = '<p>0</p><!--1--><p>1</p><!--2--><p>2</p><p>3</p><p>4</p><p>5</p>'
+    mount(parent, element)
+
+    await sleep()
+
+    const expect1 = '<p>0</p><!--1--><!--1--><p>5</p>'
+    const expect2 = '<p>0</p><!--1--><!----><p>1</p><!--2--><!--2--><p>4</p><!----><!--1--><p>5</p>'
+    const expect3 =
+      '<p>0</p><!--1--><!----><p>1</p><!--2--><!----><p>2</p><p>3</p><!----><!--2--><p>4</p><!----><!--1--><p>5</p>'
 
     bool1Atom(ctx, false)
     bool2Atom(ctx, false)
-    expect(element.innerHTML).toEqual(expect1)
-
-    bool1Atom(ctx, false)
-    bool2Atom(ctx, true)
-    expect(element.innerHTML).toEqual(expect1)
-
-    bool1Atom(ctx, true)
-    bool2Atom(ctx, false)
-    expect(element.innerHTML).toEqual(expect2)
-
-    bool1Atom(ctx, true)
-    bool2Atom(ctx, true)
-    expect(element.innerHTML).toEqual(expect3)
-
-    bool1Atom(ctx, true)
-    bool2Atom(ctx, false)
-    expect(element.innerHTML).toEqual(expect2)
+    expect(element.innerHTML).toBe(expect1)
 
     bool1Atom(ctx, false)
     bool2Atom(ctx, true)
-    expect(element.innerHTML).toEqual(expect1)
+    expect(element.innerHTML).toBe(expect1)
+
+    bool1Atom(ctx, true)
+    bool2Atom(ctx, false)
+    expect(element.innerHTML).toBe(expect2)
+
+    bool1Atom(ctx, true)
+    bool2Atom(ctx, true)
+    expect(element.innerHTML).toBe(expect3)
+
+    bool1Atom(ctx, true)
+    bool2Atom(ctx, false)
+    expect(element.innerHTML).toBe(expect2)
+
+    bool1Atom(ctx, true)
+    bool2Atom(ctx, true)
+    expect(element.innerHTML).toBe(expect3)
+
+    bool1Atom(ctx, false)
+    bool2Atom(ctx, true)
+    expect(element.innerHTML).toBe(expect1)
 
     bool1Atom(ctx, false)
     bool2Atom(ctx, false)
-    expect(element.innerHTML).toEqual(expect1)
+    expect(element.innerHTML).toBe(expect1)
   }),
 )
 
@@ -775,7 +803,52 @@ it(
 
     inputState(ctx, '43')
 
-    expect(input.value).toEqual('43')
-    expect(testSvg.innerHTML).toEqual('<path d="M 10 10 H 100"></path>')
+    expect(input.value).toBe('43')
+    expect(testSvg.innerHTML).toBe('<path d="M 10 10 H 100"></path>')
+  }),
+)
+
+it(
+  'dynamic atom fragment',
+  setup((ctx, h, hf, mount, parent) => {
+    const child = atom<JSX.HTMLAttributes['children']>(<span />, 'test')
+
+    const container = <div>{child}</div>
+    mount(parent, container)
+
+    expect(container.outerHTML).toBe('<div><!--test--><span></span><!--test--></div>')
+
+    child(ctx, () => atom('child atom', 'test.child'))
+    expect(container.outerHTML).toBe('<div><!--test--><!--test.child-->child atom<!--test.child--><!--test--></div>')
+  }),
+)
+
+it(
+  'linked list',
+  setup((ctx, h, hf, mount, parent) => {
+    const a = atom(true)
+    const list = reatomLinkedList((ctx, value: string) => (
+      <>
+        <span>{value}</span>
+        {atom((ctx) => (ctx.spy(a) ? <a /> : <br />), 'test')}
+      </>
+    ))
+    list.create(ctx, '1')
+
+    const container = <div>{list}</div>
+    mount(parent, container)
+
+    expect(container.outerHTML).toBe('<div><!----><span>1</span><!--test--><a></a><!--test--><!----></div>')
+
+    a(ctx, false)
+    expect(container.outerHTML).toBe('<div><!----><span>1</span><!--test--><br><!--test--><!----></div>')
+
+    const node = list.create(ctx, '2')
+    expect(container.outerHTML).toBe(
+      '<div><!----><span>1</span><!--test--><br><!--test--><!----><!----><span>2</span><!--test--><br><!--test--><!----></div>',
+    )
+
+    list.remove(ctx, node)
+    expect(container.outerHTML).toBe('<div><!----><span>1</span><!--test--><br><!--test--><!----></div>')
   }),
 )