+++ packages/url/src/index.test.ts
@@ -112,3 +112,24 @@ it('SearchParamsAtom.lens subpath', () => {
   urlAtom.go(ctx, '/results/some')
   expect(ctx.get(urlAtom).href).toEqual('http://example.com/results/some')
 })
+
+test('SearchParamsAtom remove query from url', () => {
+  const ctx = createTestCtx()
+
+  setupUrlAtomSettings(ctx, () => new URL('http://example.com'))
+
+  const testAtom = atom<number | undefined>(undefined).pipe(
+    withSearchParamsPersist('test', {
+      parse: (value) => value ===  undefined ? undefined : Number(value),
+      serialize: (value) => value === undefined ? undefined : value.toString(),
+    }),
+  )
+
+  ctx.subscribeTrack(testAtom)
+
+  urlAtom.go(ctx, '/results?test=2')
+  assert.is(ctx.get(testAtom), 2)
+
+  testAtom(ctx, undefined)
+  assert.is(ctx.get(urlAtom).href, 'http://example.com/results')
+})